% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{get_article}
\alias{get_article}
\title{Get SRF Articles}
\usage{
get_article(url, table_name = "SRFArticles", browser = "firefox", port = 5443L)
}
\arguments{
\item{url}{URL to the article}

\item{table_name}{name of the table}

\item{browser}{browser to use}

\item{port}{port for silenium}
}
\value{
creates an SQL table
}
\description{
Get SRF Articles
}
\examples{
get_article("https://www.srf.ch/news/schweiz/lawinengefahr-in-der-schweiz-lawinenprognostiker-zurzeit-ist-die-lawinensituation-kritisch")
}
